+++ toshi_hazard_store/model/attributes/__init__.py
@@ -6,4 +6,8 @@ from .attributes import (
     PickleAttribute,
 )
 from .enum_attribute import EnumAttribute
-from .enum_constrained_attribute import EnumConstrainedAttribute
+from .enum_constrained_attribute import (
+    EnumConstrainedFloatAttribute,
+    EnumConstrainedIntegerAttribute,
+    EnumConstrainedUnicodeAttribute,
+)