+++ toshi_hazard_store/model/attributes/enum_attribute.py
@@ -1,4 +1,5 @@
 """This module defines a custom enum attribute."""
+import logging
 from enum import Enum
 from typing import Any, Type, TypeVar
 
@@ -6,6 +7,9 @@ from pynamodb.attributes import Attribute
 from pynamodb.constants import STRING
 
 T = TypeVar("T", bound=Enum)
+_T = TypeVar('_T')
+
+log = logging.getLogger(__name__)
 
 
 class EnumAttribute(Attribute[T]):
@@ -36,14 +40,25 @@ class EnumAttribute(Attribute[T]):
         super().__init__(**kwargs)
         self.enum_type = enum_type
 
-    def deserialize(self, name: str) -> Type[T]:
+    def deserialize(self, value: str) -> Type[T]:
+        log.info(f'user deserialize value {value}')
         try:
-            return getattr(self.enum_type, name)
-            return name
-        except (AttributeError):
-            raise ValueError(f'stored value {name} must be a value of  {self.enum_type}.')
+            val = self.enum_type[value]  # getattr(self.enum_type, value)
+            log.info(f'enum: {val}')
+            # return val
+            return super().deserialize(val)
+        except (AttributeError, KeyError):
+            raise ValueError(f'stored value {value} must be a member of  {self.enum_type}.')
 
-    def serialize(self, instance: Type[T]) -> str:
-        if isinstance(instance, self.enum_type):
-            return instance.name
-        raise ValueError(f'value {instance} must be a {self.enum_type} or an instance.')
+    def serialize(self, value: Type[T]) -> str:
+        log.info(f'user serialize value {value}')
+        if isinstance(value, self.enum_type):
+            print(f'serialize value {value}')
+            return super().serialize(value.name)
+        else:
+            try:
+                assert self.enum_type(value)  # CBC MARKS
+                return super().serialize(value)
+            except (Exception) as err:
+                print(err)
+                raise ValueError(f'value {value} must be a member of {self.enum_type}.')