+++ toshi_hazard_store/model/attributes/enum_constrained_attribute.py
@@ -1,21 +1,54 @@
 """This module defines some custom enum attributes."""
-
+import logging
 from enum import Enum
-from typing import Any, Type, TypeVar, Union
+from typing import Any, Optional, Type, TypeVar, Union
 
 from pynamodb.attributes import Attribute
 from pynamodb.constants import NUMBER, STRING
 
 T = TypeVar("T", bound=Enum)
+_T = TypeVar('_T')
+
+log = logging.getLogger(__name__)
 
 
-class EnumConstrainedAttribute(Attribute[T]):
+class EnumConstrainedAttributeMixin:
+
+    attr_type = STRING
+    value_type = Any
+    enum_type: Any
+
+    def _validate_enum(self, enum_type, *args):
+        if not all(isinstance(e.value, self.value_type) for e in self.enum_type):
+            raise TypeError(
+                f"Enumeration '{self.enum_type}' values must be all {self.value_type}",
+            )
+
+    def deserialize(self, value: Any) -> Union[str, float, int]:
+        try:
+            assert self.enum_type(value)
+            return value
+        except (ValueError):
+            raise ValueError(f'value {value} must be a member of {self.enum_type}')
+
+    def serialize(self, value: Any) -> str:
+        try:
+            if isinstance(value, self.enum_type):
+                log.info(f'user passed in enum type {value} {str(value.value)}')  # type: ignore
+                return str(value.value)
+            # if not isinstance(value, self.value_type):
+            #     raise ValueError(f'value {value} must be a member of {self.enum_type}')
+            self.enum_type(value)
+        except (ValueError) as err:
+            raise err
+        return str(value)
+
+
+class EnumConstrainedUnicodeAttribute(EnumConstrainedAttributeMixin, Attribute[T]):
     """
-    Stores strings or numbers that are values of the supplied Enum as DynamoDB strings.
+    Stores values of the supplied Unicode Enum as DynamoDB STRING types.
 
     Useful where you have values in an existing table field and you want retrofit Enum validation.
-
-    Otherwise, consider using the EnumAttribute class below and
 
     >>> from enum import Enum
     >>> from pynamodb.models import Model
@@ -31,59 +64,15 @@ class EnumConstrainedAttribute(Attribute[T]):
     >>>
     """
 
-    def __init__(self, enum_type: Type[T], attr_type: str = STRING, **kwargs: Any) -> None:
-        """
-        :param enum_type: The type of the enum
-        """
-        assert attr_type in [STRING, NUMBER]
-        self.attr_type = attr_type
-
-        super().__init__(**kwargs)
-        self.enum_type = enum_type
-
-        self.valid_types = (str,) if attr_type == STRING else (int, float)
-        if not all(isinstance(e.value, self.valid_types) for e in self.enum_type):
-            raise TypeError(
-                f"Enumeration '{self.enum_type}' values must be all {self.valid_types}",
-            )
-
-    def deserialize(self, value: Union[str, float, int]) -> Union[str, float, int]:
-        try:
-            assert self.enum_type(value)
-            return value
-        except (ValueError):
-            raise ValueError(f'value {value} must be a member of {self.enum_type}')
-
-    def serialize(self, value: Union[str, float, int]) -> Union[str, float, int]:
-        try:
-            if not isinstance(value, self.valid_types):
-                raise ValueError(f'value {value} must be a member of {self.enum_type}')
-            self.enum_type(value)
-        except (ValueError) as err:
-            raise err
-        return value
-
-
-class EnumAttribute(Attribute[T]):
-    """
-    Stores names of the supplied Enum as DynamoDB strings.
-
-    >>> from enum import Enum
-    >>>
-    >>> from pynamodb.models import Model
-    >>>
-    >>> class ShakeFlavor(Enum):
-    >>>   VANILLA = 0.1
-    >>>   MINT = 1.22
-    >>>
-    >>> class Shake(Model):
-    >>>   flavor = EnumAttribute(ShakeFlavor)
-    >>>
-    >>> modelB = Shake(flavor=ShakeFlavor.MINT)
-    >>> assert modelB.flavor == ShakeFlavor.MINT
-    """
-
     attr_type = STRING
+    value_type = str
+
+    def __set__(self, instance: Any, value: Optional[T]) -> None:
+        if isinstance(value, self.enum_type):
+            log.info(f'user __set__ enum type {value} {str(value.value)}')
+            super().__set__(instance, value.value)
+        else:
+            super().__set__(instance, value)
 
     def __init__(self, enum_type: Type[T], **kwargs: Any) -> None:
         """
@@ -91,40 +80,18 @@ class EnumAttribute(Attribute[T]):
         """
         super().__init__(**kwargs)
         self.enum_type = enum_type
+        super()._validate_enum(self.enum_type, self.value_type)
 
-    def deserialize(self, name: str) -> Type[T]:
-        try:
-            return getattr(self.enum_type, name)
-            return name
-        except (AttributeError):
-            raise ValueError(f'stored value {name} must be a value of  {self.enum_type}.')
+    def deserialize(self, value: Union[float, int]) -> str:
+        return str(super().deserialize(value))
 
-    def serialize(self, instance: Type[T]) -> str:
-        if isinstance(instance, self.enum_type):
-            return instance.name
-        raise ValueError(f'value {instance} must be a {self.enum_type} or an instance.')
+    def serialize(self, value: Union[float, int]) -> str:
+        return super().serialize(value)
 
 
-class NumberEnumAttribute(Attribute[T]):
-    """
-    Stores names of the supplied Enum as DynamoDB number.
-
-    >>> from enum import Enum
-    >>>
-    >>> from pynamodb.models import Model
-    >>>
-    >>> class ShakeFlavor(Enum):
-    >>>   VANILLA = 0.1
-    >>>   MINT = 1.22
-    >>>
-    >>> class Shake(Model):
-    >>>   flavor = EnumAttribute(ShakeFlavor)
-    >>>
-    >>> modelB = Shake(flavor=ShakeFlavor.MINT)
-    >>> assert modelB.flavor == ShakeFlavor.MINT
-    """
-
+class EnumConstrainedIntegerAttribute(EnumConstrainedAttributeMixin, Attribute[T]):
     attr_type = NUMBER
+    value_type = int
 
     def __init__(self, enum_type: Type[T], **kwargs: Any) -> None:
         """
@@ -132,15 +99,43 @@ class NumberEnumAttribute(Attribute[T]):
         """
         super().__init__(**kwargs)
         self.enum_type = enum_type
+        super()._validate_enum(self.enum_type, self.value_type)
 
-    def deserialize(self, name: str) -> Type[T]:
-        try:
-            return getattr(self.enum_type, name)
-            return name
-        except (AttributeError):
-            raise ValueError(f'stored value {name} must be a value of  {self.enum_type}.')
+    def __set__(self, instance: Any, value: Optional[T]) -> None:
+        if isinstance(value, self.enum_type):
+            log.info(f'user __set__ enum type {value} {str(value.value)}')
+            super().__set__(instance, value.value)
+        else:
+            super().__set__(instance, value)
 
-    def serialize(self, instance: Type[T]) -> str:
-        if isinstance(instance, self.enum_type):
-            return instance.name
-        raise ValueError(f'value {instance} must be a {self.enum_type} or an instance.')
+    def deserialize(self, value: Union[float, int]) -> int:
+        return int(super().deserialize(int(value)))
+
+    def serialize(self, value: Union[float, int]) -> str:
+        return super().serialize(int(value))
+
+
+class EnumConstrainedFloatAttribute(EnumConstrainedAttributeMixin, Attribute[T]):
+    attr_type = NUMBER
+    value_type = float
+
+    def __init__(self, enum_type: Type[T], **kwargs: Any) -> None:
+        """
+        :param enum_type: The type of the enum
+        """
+        super().__init__(**kwargs)
+        self.enum_type = enum_type
+        super()._validate_enum(self.enum_type, self.value_type)
+
+    def __set__(self, instance: Any, value: Optional[T]) -> None:
+        if isinstance(value, self.enum_type):
+            log.info(f'user __set__ enum type {value} {str(value.value)}')
+            super().__set__(instance, value.value)
+        else:
+            super().__set__(instance, value)
+
+    def deserialize(self, value: Union[float, int]) -> float:
+        return float(super().deserialize(float(value)))
+
+    def serialize(self, value: Union[float, int]) -> str:
+        return super().serialize(float(value))